Fix destination phone number in sample data (#22)
I was validating the phone number format and notice probably a typo in sample data. If the destination number is (99) 9346-8278, it should have the digit 9 (nine) at the left: (99) **9**9346-8278 or the call would not be completed. Am I right?

Since 2016, the digit 9 (nine) was added to the left of the current mobile numbers.
Sources:
- http://www.teleco.com.br/num_cel.asp
- http://www.teleco.com.br/num.asp
- http://www.anatel.gov.br/Portal/verificaDocumentos/documento.asp?null&filtro=1&documentoPath=biblioteca/resolucao/1998/anexo_res_86_1998.pdf